+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 + Real-time network and disk usage [#68](https://github.com/lvlup-pro/spawn-frontend-material/issues/68)
 
+### Fixes
+
++ Force logout after 401 or 403 error [#69](https://github.com/lvlup-pro/spawn-frontend-material/issues/69)
+
 ## 3.0.0-alpha.9 (23.02.2017)
 
 ### Features